+++ attributes/phantomjs.rb
@@ -1,5 +1,0 @@
-default['selenium']['phantomjs']['webdriver'] = "#{node['ipaddress']}:8910"
-default['selenium']['phantomjs']['webdriverSeleniumGridHub'] = "http://#{node['ipaddress']}:4444"
-default['selenium']['phantomjs']['username'] = nil
-default['selenium']['phantomjs']['password'] = nil
-default['selenium']['phantomjs']['domain'] = nil