+++ libraries/default.rb
@@ -1,29 +1,10 @@
-CHROME ||= 'chrome'
-FIREFOX ||= 'firefox'
-IE ||= 'internet explorer'
-SAFARI ||= 'safari'
 
 private
-
-def selenium_browser?(browser, capabilities)
-  return false if capabilities.nil?
-  capabilities.each do |capability|
-    return true if capability[:browserName] == browser
-  end
-  false
-end
 
 def selenium_java_exec
   java = platform_family?('windows') ? node['selenium']['windows']['java'] : node['selenium']['linux']['java']
   validate_exec(%("#{java}" -version))
   java
-end
-
-def selenium_phantomjs_exec
-  phantomjs =
-    platform_family?('windows') ? node['selenium']['windows']['phantomjs'] : node['selenium']['linux']['phantomjs']
-  validate_exec(%("#{phantomjs}" -v))
-  phantomjs
 end
 
 def validate_exec(cmd)
@@ -36,12 +17,8 @@ def selenium_home
   platform_family?('windows') ? node['selenium']['windows']['home'] : node['selenium']['linux']['home']
 end
 
-def selenium_server_standalone
+def selenium_jar_link
   "#{selenium_home}/server/selenium-server-standalone.jar"
-end
-
-def selenium_version(version)
-  version.slice(0, version.rindex('.'))
 end
 
 def windows_service(name, exec, args)
@@ -172,12 +149,4 @@ end
 
 def mac_domain(name)
   "org.seleniumhq.#{name}"
-end
-
-# Fixes #15: Drivers not copied to /selenium/drivers/ folders on Windows 7
-def powershell_version
-  out = shell_out!('powershell.exe $host.version.major')
-  out.to_i
-rescue # powershell not installed
-  -1
 end