+++ recipes/node.rb
@@ -1,4 +1,4 @@
-selenium_node node['selenium']['node']['service_name'] do
+selenium_node node['selenium']['node']['servicename'] do
   host node['selenium']['node']['host']
   port node['selenium']['node']['port']
   jvm_args node['selenium']['node']['jvm_args']