+++ test/fixtures/cookbooks/selenium_test/metadata.rb
@@ -6,8 +6,7 @@ description 'Tests Selenium'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.1.0'
 
-depends 'aws'
-depends 'java'
+depends 'java_se', '~> 8.0'
 depends 'firefox'
 depends 'chrome'
 depends 'safari'