+++ test/fixtures/cookbooks/selenium_test/recipes/safaridriver.rb
@@ -1,4 +1,0 @@
-node.set['selenium']['safaridriver_username'] = 'vagrant'
-node.set['selenium']['safaridriver_password'] = 'vagrant'
-
-include_recipe 'selenium::safaridriver'